refactor parameter/tag parsing
