+++ src/main/java/de/taimos/pipeline/aws/cloudformation/ParameterProvider.java
@@ -1,0 +1,12 @@
+package de.taimos.pipeline.aws.cloudformation;
+
+import java.io.File;
+
+public interface ParameterProvider {
+    String[] getKeepParams();
+
+    File getParamsFile();
+
+    String[] getParams();
+
+}
