+++ src/main/java/de/taimos/pipeline/aws/cloudformation/stacksets/StackSetOperationFailedException.java
@@ -2,13 +2,13 @@ package de.taimos.pipeline.aws.cloudformation.stacksets;
 
 public class StackSetOperationFailedException extends IllegalStateException {
 
-    private final String operationId;
+	private final String operationId;
 
-    public StackSetOperationFailedException(String operationId) {
-        this.operationId = operationId;
-    }
+	public StackSetOperationFailedException(String operationId) {
+		this.operationId = operationId;
+	}
 
-    public String getOperationId() {
-        return operationId;
-    }
+	public String getOperationId() {
+		return operationId;
+	}
 }
