--- src/test/java/de/taimos/pipeline/aws/cloudformation/parser/ParameterParserTests.java
@@ -1,4 +1,0 @@
-package de.taimos.pipeline.aws.cloudformation.parser;
-
-public class ParameterParserTests {
-}
