+++ src/test/java/de/taimos/pipeline/aws/cloudformation/stacksets/CloudFormationStackSetTest.java
@@ -11,257 +11,256 @@ import org.mockito.Mockito;
 
 import java.util.Collections;
 import java.util.UUID;
-import java.util.concurrent.ExecutionException;
 
 public class CloudFormationStackSetTest {
 
-    private TaskListener listener;
-    private AmazonCloudFormation client;
-    private CloudFormationStackSet stackSet;
+	private TaskListener listener;
+	private AmazonCloudFormation client;
+	private CloudFormationStackSet stackSet;
 
-    @Before
-    public void setup() {
-        listener = Mockito.mock(TaskListener.class);
-        Mockito.when(listener.getLogger()).thenReturn(System.out);
-        client = Mockito.mock(AmazonCloudFormation.class);
-        stackSet = new CloudFormationStackSet(client, "foo", listener);
-    }
+	@Before
+	public void setup() {
+		listener = Mockito.mock(TaskListener.class);
+		Mockito.when(listener.getLogger()).thenReturn(System.out);
+		client = Mockito.mock(AmazonCloudFormation.class);
+		stackSet = new CloudFormationStackSet(client, "foo", listener);
+	}
 
-    @Test
-    public void stackSetExists() {
-        Mockito.when(client.describeStackSet(Mockito.any(DescribeStackSetRequest.class)))
-                .thenReturn(new DescribeStackSetResult()
-                        .withStackSet(new StackSet())
-                );
+	@Test
+	public void stackSetExists() {
+		Mockito.when(client.describeStackSet(Mockito.any(DescribeStackSetRequest.class)))
+				.thenReturn(new DescribeStackSetResult()
+						.withStackSet(new StackSet())
+				);
 
-        Assertions.assertThat(stackSet.exists()).isTrue();
-    }
+		Assertions.assertThat(stackSet.exists()).isTrue();
+	}
 
-    @Test
-    public void stackSetDoesNotExists() {
-        AmazonCloudFormationException ex = new AmazonCloudFormationException("stack set does not exist");
-        ex.setErrorCode("StackSetNotFoundException");
-        Mockito.when(client.describeStackSet(Mockito.any(DescribeStackSetRequest.class)))
-                .thenThrow(ex);
+	@Test
+	public void stackSetDoesNotExists() {
+		AmazonCloudFormationException ex = new AmazonCloudFormationException("stack set does not exist");
+		ex.setErrorCode("StackSetNotFoundException");
+		Mockito.when(client.describeStackSet(Mockito.any(DescribeStackSetRequest.class)))
+				.thenThrow(ex);
 
-        Assertions.assertThat(stackSet.exists()).isFalse();
-    }
+		Assertions.assertThat(stackSet.exists()).isFalse();
+	}
 
-    @Test(expected = AmazonCloudFormationException.class)
-    public void stackSetExistsError() {
-        AmazonCloudFormationException ex = new AmazonCloudFormationException("stack set does not exist");
-        Mockito.when(client.describeStackSet(Mockito.any(DescribeStackSetRequest.class)))
-                .thenThrow(ex);
+	@Test(expected = AmazonCloudFormationException.class)
+	public void stackSetExistsError() {
+		AmazonCloudFormationException ex = new AmazonCloudFormationException("stack set does not exist");
+		Mockito.when(client.describeStackSet(Mockito.any(DescribeStackSetRequest.class)))
+				.thenThrow(ex);
 
-        stackSet.exists();
-    }
+		stackSet.exists();
+	}
 
-    @Test
-    public void createTemplateBody() {
-        CreateStackSetResult expected = new CreateStackSetResult();
-        Mockito.when(client.createStackSet(Mockito.any(CreateStackSetRequest.class)))
-                .thenReturn(expected);
+	@Test
+	public void createTemplateBody() {
+		CreateStackSetResult expected = new CreateStackSetResult();
+		Mockito.when(client.createStackSet(Mockito.any(CreateStackSetRequest.class)))
+				.thenReturn(expected);
 
-        Parameter parameter1 = new Parameter()
-                .withParameterKey("foo")
-                .withParameterValue("bar");
-        Tag tag1 = new Tag()
-                .withKey("bar")
-                .withValue("baz");
+		Parameter parameter1 = new Parameter()
+				.withParameterKey("foo")
+				.withParameterValue("bar");
+		Tag tag1 = new Tag()
+				.withKey("bar")
+				.withValue("baz");
 
-        CreateStackSetResult result = stackSet.create("body", null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
-        Assertions.assertThat(result).isSameAs(expected);
-        ArgumentCaptor<CreateStackSetRequest> captor = ArgumentCaptor.forClass(CreateStackSetRequest.class);
-        Mockito.verify(client).createStackSet(captor.capture());
-        Assertions.assertThat(captor.getValue()).isEqualTo(new CreateStackSetRequest()
-                .withStackSetName("foo")
-                .withCapabilities(Capability.values())
-                .withParameters(parameter1)
-                .withTags(tag1)
-                .withTemplateBody("body")
-        );
-    }
+		CreateStackSetResult result = stackSet.create("body", null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
+		Assertions.assertThat(result).isSameAs(expected);
+		ArgumentCaptor<CreateStackSetRequest> captor = ArgumentCaptor.forClass(CreateStackSetRequest.class);
+		Mockito.verify(client).createStackSet(captor.capture());
+		Assertions.assertThat(captor.getValue()).isEqualTo(new CreateStackSetRequest()
+				.withStackSetName("foo")
+				.withCapabilities(Capability.values())
+				.withParameters(parameter1)
+				.withTags(tag1)
+				.withTemplateBody("body")
+		);
+	}
 
-    @Test
-    public void createTemplateUrl() {
-        CreateStackSetResult expected = new CreateStackSetResult();
-        Mockito.when(client.createStackSet(Mockito.any(CreateStackSetRequest.class)))
-                .thenReturn(expected);
+	@Test
+	public void createTemplateUrl() {
+		CreateStackSetResult expected = new CreateStackSetResult();
+		Mockito.when(client.createStackSet(Mockito.any(CreateStackSetRequest.class)))
+				.thenReturn(expected);
 
-        Parameter parameter1 = new Parameter()
-                .withParameterKey("foo")
-                .withParameterValue("bar");
-        Tag tag1 = new Tag()
-                .withKey("bar")
-                .withValue("baz");
+		Parameter parameter1 = new Parameter()
+				.withParameterKey("foo")
+				.withParameterValue("bar");
+		Tag tag1 = new Tag()
+				.withKey("bar")
+				.withValue("baz");
 
-        CreateStackSetResult result = stackSet.create(null, "url", Collections.singletonList(parameter1), Collections.singletonList(tag1));
-        Assertions.assertThat(result).isSameAs(expected);
-        ArgumentCaptor<CreateStackSetRequest> captor = ArgumentCaptor.forClass(CreateStackSetRequest.class);
-        Mockito.verify(client).createStackSet(captor.capture());
-        Assertions.assertThat(captor.getValue()).isEqualTo(new CreateStackSetRequest()
-                .withStackSetName("foo")
-                .withCapabilities(Capability.values())
-                .withParameters(parameter1)
-                .withTags(tag1)
-                .withTemplateURL("url")
-        );
-    }
+		CreateStackSetResult result = stackSet.create(null, "url", Collections.singletonList(parameter1), Collections.singletonList(tag1));
+		Assertions.assertThat(result).isSameAs(expected);
+		ArgumentCaptor<CreateStackSetRequest> captor = ArgumentCaptor.forClass(CreateStackSetRequest.class);
+		Mockito.verify(client).createStackSet(captor.capture());
+		Assertions.assertThat(captor.getValue()).isEqualTo(new CreateStackSetRequest()
+				.withStackSetName("foo")
+				.withCapabilities(Capability.values())
+				.withParameters(parameter1)
+				.withTags(tag1)
+				.withTemplateURL("url")
+		);
+	}
 
-    @Test(expected = IllegalArgumentException.class)
-    public void createNoTemplate() {
-        CreateStackSetResult expected = new CreateStackSetResult();
-        Mockito.when(client.createStackSet(Mockito.any(CreateStackSetRequest.class)))
-                .thenReturn(expected);
+	@Test(expected = IllegalArgumentException.class)
+	public void createNoTemplate() {
+		CreateStackSetResult expected = new CreateStackSetResult();
+		Mockito.when(client.createStackSet(Mockito.any(CreateStackSetRequest.class)))
+				.thenReturn(expected);
 
-        Parameter parameter1 = new Parameter()
-                .withParameterKey("foo")
-                .withParameterValue("bar");
-        Tag tag1 = new Tag()
-                .withKey("bar")
-                .withValue("baz");
+		Parameter parameter1 = new Parameter()
+				.withParameterKey("foo")
+				.withParameterValue("bar");
+		Tag tag1 = new Tag()
+				.withKey("bar")
+				.withValue("baz");
 
-        stackSet.create(null, null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
-    }
+		stackSet.create(null, null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
+	}
 
-    @Test
-    public void updateTemplateBody() {
-        UpdateStackSetResult expected = new UpdateStackSetResult();
-        Mockito.when(client.updateStackSet(Mockito.any(UpdateStackSetRequest.class)))
-                .thenReturn(expected);
+	@Test
+	public void updateTemplateBody() {
+		UpdateStackSetResult expected = new UpdateStackSetResult();
+		Mockito.when(client.updateStackSet(Mockito.any(UpdateStackSetRequest.class)))
+				.thenReturn(expected);
 
-        Parameter parameter1 = new Parameter()
-                .withParameterKey("foo")
-                .withParameterValue("bar");
-        Tag tag1 = new Tag()
-                .withKey("bar")
-                .withValue("baz");
+		Parameter parameter1 = new Parameter()
+				.withParameterKey("foo")
+				.withParameterValue("bar");
+		Tag tag1 = new Tag()
+				.withKey("bar")
+				.withValue("baz");
 
-        UpdateStackSetResult result = stackSet.update("body", null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
-        Assertions.assertThat(result).isSameAs(expected);
-        ArgumentCaptor<UpdateStackSetRequest> captor = ArgumentCaptor.forClass(UpdateStackSetRequest.class);
-        Mockito.verify(client).updateStackSet(captor.capture());
-        Assertions.assertThat(captor.getValue()).isEqualTo(new UpdateStackSetRequest()
-                .withStackSetName("foo")
-                .withCapabilities(Capability.values())
-                .withParameters(parameter1)
-                .withTags(tag1)
-                .withTemplateBody("body")
-        );
-    }
+		UpdateStackSetResult result = stackSet.update("body", null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
+		Assertions.assertThat(result).isSameAs(expected);
+		ArgumentCaptor<UpdateStackSetRequest> captor = ArgumentCaptor.forClass(UpdateStackSetRequest.class);
+		Mockito.verify(client).updateStackSet(captor.capture());
+		Assertions.assertThat(captor.getValue()).isEqualTo(new UpdateStackSetRequest()
+				.withStackSetName("foo")
+				.withCapabilities(Capability.values())
+				.withParameters(parameter1)
+				.withTags(tag1)
+				.withTemplateBody("body")
+		);
+	}
 
-    @Test
-    public void updateTemplateUrl() {
-        UpdateStackSetResult expected = new UpdateStackSetResult();
-        Mockito.when(client.updateStackSet(Mockito.any(UpdateStackSetRequest.class)))
-                .thenReturn(expected);
+	@Test
+	public void updateTemplateUrl() {
+		UpdateStackSetResult expected = new UpdateStackSetResult();
+		Mockito.when(client.updateStackSet(Mockito.any(UpdateStackSetRequest.class)))
+				.thenReturn(expected);
 
-        Parameter parameter1 = new Parameter()
-                .withParameterKey("foo")
-                .withParameterValue("bar");
-        Tag tag1 = new Tag()
-                .withKey("bar")
-                .withValue("baz");
+		Parameter parameter1 = new Parameter()
+				.withParameterKey("foo")
+				.withParameterValue("bar");
+		Tag tag1 = new Tag()
+				.withKey("bar")
+				.withValue("baz");
 
-        UpdateStackSetResult result = stackSet.update(null, "url", Collections.singletonList(parameter1), Collections.singletonList(tag1));
-        Assertions.assertThat(result).isSameAs(expected);
-        ArgumentCaptor<UpdateStackSetRequest> captor = ArgumentCaptor.forClass(UpdateStackSetRequest.class);
-        Mockito.verify(client).updateStackSet(captor.capture());
-        Assertions.assertThat(captor.getValue()).isEqualTo(new UpdateStackSetRequest()
-                .withStackSetName("foo")
-                .withCapabilities(Capability.values())
-                .withParameters(parameter1)
-                .withTags(tag1)
-                .withTemplateURL("url")
-        );
-    }
+		UpdateStackSetResult result = stackSet.update(null, "url", Collections.singletonList(parameter1), Collections.singletonList(tag1));
+		Assertions.assertThat(result).isSameAs(expected);
+		ArgumentCaptor<UpdateStackSetRequest> captor = ArgumentCaptor.forClass(UpdateStackSetRequest.class);
+		Mockito.verify(client).updateStackSet(captor.capture());
+		Assertions.assertThat(captor.getValue()).isEqualTo(new UpdateStackSetRequest()
+				.withStackSetName("foo")
+				.withCapabilities(Capability.values())
+				.withParameters(parameter1)
+				.withTags(tag1)
+				.withTemplateURL("url")
+		);
+	}
 
-    @Test
-    public void updateTemplateKeepPrevious() {
-        UpdateStackSetResult expected = new UpdateStackSetResult();
-        Mockito.when(client.updateStackSet(Mockito.any(UpdateStackSetRequest.class)))
-                .thenReturn(expected);
+	@Test
+	public void updateTemplateKeepPrevious() {
+		UpdateStackSetResult expected = new UpdateStackSetResult();
+		Mockito.when(client.updateStackSet(Mockito.any(UpdateStackSetRequest.class)))
+				.thenReturn(expected);
 
-        Parameter parameter1 = new Parameter()
-                .withParameterKey("foo")
-                .withParameterValue("bar");
-        Tag tag1 = new Tag()
-                .withKey("bar")
-                .withValue("baz");
+		Parameter parameter1 = new Parameter()
+				.withParameterKey("foo")
+				.withParameterValue("bar");
+		Tag tag1 = new Tag()
+				.withKey("bar")
+				.withValue("baz");
 
-        UpdateStackSetResult result = stackSet.update(null, null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
-        Assertions.assertThat(result).isSameAs(expected);
-        ArgumentCaptor<UpdateStackSetRequest> captor = ArgumentCaptor.forClass(UpdateStackSetRequest.class);
-        Mockito.verify(client).updateStackSet(captor.capture());
-        Assertions.assertThat(captor.getValue()).isEqualTo(new UpdateStackSetRequest()
-                .withStackSetName("foo")
-                .withCapabilities(Capability.values())
-                .withParameters(parameter1)
-                .withTags(tag1)
-                .withUsePreviousTemplate(true)
-        );
-    }
+		UpdateStackSetResult result = stackSet.update(null, null, Collections.singletonList(parameter1), Collections.singletonList(tag1));
+		Assertions.assertThat(result).isSameAs(expected);
+		ArgumentCaptor<UpdateStackSetRequest> captor = ArgumentCaptor.forClass(UpdateStackSetRequest.class);
+		Mockito.verify(client).updateStackSet(captor.capture());
+		Assertions.assertThat(captor.getValue()).isEqualTo(new UpdateStackSetRequest()
+				.withStackSetName("foo")
+				.withCapabilities(Capability.values())
+				.withParameters(parameter1)
+				.withTags(tag1)
+				.withUsePreviousTemplate(true)
+		);
+	}
 
-    @Test
-    public void waitForStackStateStatus() throws InterruptedException {
-        Mockito.when(client.describeStackSet(new DescribeStackSetRequest()
-                .withStackSetName("foo")
-        )).thenReturn(new DescribeStackSetResult()
-                .withStackSet(new StackSet()
-                        .withStatus(StackSetStatus.ACTIVE)
-                )
-        ).thenReturn(new DescribeStackSetResult()
-                .withStackSet(new StackSet()
-                        .withStatus(StackSetStatus.DELETED)
-                )
-        );
-        stackSet.waitForStackState(StackSetStatus.DELETED, 5);
+	@Test
+	public void waitForStackStateStatus() throws InterruptedException {
+		Mockito.when(client.describeStackSet(new DescribeStackSetRequest()
+				.withStackSetName("foo")
+		)).thenReturn(new DescribeStackSetResult()
+				.withStackSet(new StackSet()
+						.withStatus(StackSetStatus.ACTIVE)
+				)
+		).thenReturn(new DescribeStackSetResult()
+				.withStackSet(new StackSet()
+						.withStatus(StackSetStatus.DELETED)
+				)
+		);
+		stackSet.waitForStackState(StackSetStatus.DELETED, 5);
 
-        Mockito.verify(client, Mockito.atLeast(2))
-                .describeStackSet(Mockito.any(DescribeStackSetRequest.class));
-    }
+		Mockito.verify(client, Mockito.atLeast(2))
+				.describeStackSet(Mockito.any(DescribeStackSetRequest.class));
+	}
 
-    @Test
-    public void waitForOperationToComplete() throws InterruptedException {
-        String operationId = UUID.randomUUID().toString();
-        Mockito.when(client.describeStackSetOperation(new DescribeStackSetOperationRequest()
-                .withStackSetName("foo")
-                .withOperationId(operationId)
-        )).thenReturn(new DescribeStackSetOperationResult()
-                .withStackSetOperation(new StackSetOperation()
-                        .withStatus(StackSetOperationStatus.RUNNING)
-                )
-        ).thenReturn(new DescribeStackSetOperationResult()
-                        .withStackSetOperation(new StackSetOperation()
-                                .withStatus(StackSetOperationStatus.SUCCEEDED)
-                        )
-        );
-        stackSet.waitForOperationToComplete(operationId, 5);
-    }
+	@Test
+	public void waitForOperationToComplete() throws InterruptedException {
+		String operationId = UUID.randomUUID().toString();
+		Mockito.when(client.describeStackSetOperation(new DescribeStackSetOperationRequest()
+				.withStackSetName("foo")
+				.withOperationId(operationId)
+		)).thenReturn(new DescribeStackSetOperationResult()
+				.withStackSetOperation(new StackSetOperation()
+						.withStatus(StackSetOperationStatus.RUNNING)
+				)
+		).thenReturn(new DescribeStackSetOperationResult()
+				.withStackSetOperation(new StackSetOperation()
+						.withStatus(StackSetOperationStatus.SUCCEEDED)
+				)
+		);
+		stackSet.waitForOperationToComplete(operationId, 5);
+	}
 
-    @Test(expected = StackSetOperationFailedException.class)
-    public void waitForOperationToCompleteFailure() throws InterruptedException {
-        String operationId = UUID.randomUUID().toString();
-        Mockito.when(client.describeStackSetOperation(new DescribeStackSetOperationRequest()
-                .withStackSetName("foo")
-                .withOperationId(operationId)
-        )).thenReturn(new DescribeStackSetOperationResult()
-                .withStackSetOperation(new StackSetOperation()
-                        .withStatus(StackSetOperationStatus.RUNNING)
-                )
-        ).thenReturn(new DescribeStackSetOperationResult()
-                .withStackSetOperation(new StackSetOperation()
-                        .withStatus(StackSetOperationStatus.FAILED)
-                )
-        );
-        stackSet.waitForOperationToComplete(operationId, 5);
-    }
+	@Test(expected = StackSetOperationFailedException.class)
+	public void waitForOperationToCompleteFailure() throws InterruptedException {
+		String operationId = UUID.randomUUID().toString();
+		Mockito.when(client.describeStackSetOperation(new DescribeStackSetOperationRequest()
+				.withStackSetName("foo")
+				.withOperationId(operationId)
+		)).thenReturn(new DescribeStackSetOperationResult()
+				.withStackSetOperation(new StackSetOperation()
+						.withStatus(StackSetOperationStatus.RUNNING)
+				)
+		).thenReturn(new DescribeStackSetOperationResult()
+				.withStackSetOperation(new StackSetOperation()
+						.withStatus(StackSetOperationStatus.FAILED)
+				)
+		);
+		stackSet.waitForOperationToComplete(operationId, 5);
+	}
 
-    @Test
-    public void delete() {
-        stackSet.delete();
-        Mockito.verify(client).deleteStackSet(new DeleteStackSetRequest()
-                .withStackSetName("foo")
-        );
-    }
+	@Test
+	public void delete() {
+		stackSet.delete();
+		Mockito.verify(client).deleteStackSet(new DeleteStackSetRequest()
+				.withStackSetName("foo")
+		);
+	}
 }
